make event wall function to common utility function
