+++ horaservices/src/components/wonderland/event-wall/EventwallSection.jsx
@@ -15,6 +15,7 @@ import {
 } from "@/utils/eventCache";
 import "../../common/EventLazyImage.css";
 import EventwallGalleryItem from "./EventwallGalleryItem";
+import { processImagesWithHeight } from "@/utils/eventWallHelpers";
 
 const EventwallSection = ({
   userData,
@@ -37,57 +38,6 @@ const EventwallSection = ({
   const MAX_PARALLEL_UPLOADS = 5;
   let activeUploads = 0;
   let uploadQueue = [];
-
-  // Measure height of thumbnail/local image
-  function measureImageHeight(url) {
-    return new Promise((resolve) => {
-      const img = new Image();
-      img.src = url;
-      img.onload = () => resolve(img.height);
-      img.onerror = () => resolve(0);
-    });
-  }
-
-  // Reorder — Tallest image → Big block (pos 3)
-  function reorderByHeight(items) {
-    const result = [];
-
-    for (let i = 0; i < items.length; i += 6) {
-      const chunk = items.slice(i, i + 6);
-
-      if (chunk.length < 6) {
-        result.push(...chunk);
-        continue;
-      }
-
-      const tallest = [...chunk].sort((a, b) => b.height - a.height)[0];
-
-      const arranged = [];
-      chunk.forEach((img) => {
-        if (img === tallest) return;
-        arranged.push(img);
-      });
-
-      arranged.splice(3, 0, tallest);
-      result.push(...arranged);
-    }
-
-    return result;
-  }
-
-  // Measure heights + reorder
-  async function processImagesWithHeight(list) {
-    const enriched = await Promise?.all(
-      list?.map(async (item) => ({
-        ...item,
-        height: await measureImageHeight(
-          item?.postWebpUrl || item?.postUrl || item?.localPreview
-        ),
-      }))
-    );
-
-    return reorderByHeight(enriched);
-  }
 
   useEffect(() => {
     async function loadEventPosts() {
